+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Pass project path to development environment container (gorarakelyan)
+
 ## 2.0.2 Jul 7 2020
 - Make `epoch` argument optional for `Metric` artifact (gorarakelyan)
 - Add ability to automatically commit runs after exit (gorarakelyan)